インデックスバッファなどを追加しました

+++ 3DShootingGame/3DShootingGame/Lib/Graphics/Graphics.cpp
@@ -125,9 +125,9 @@
 	void Graphics::PresentParametersConfig(
 		D3DPRESENT_PARAMETERS & d3d_pp,
 		HWND window_handle,
-		float width_size,
-		float height_size,
-		float back_buffer_count,
+		UINT width_size,
+		UINT height_size,
+		UINT back_buffer_count,
 		bool windowed
 	) {
